Encapsulated invariant jagged array bit packing code in an abstract class
